# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Clerk Platform API",
    "x-logo": {
      "url": "https://clerk.com/_next/image?url=%2Fimages%2Fclerk-logo.svg&w=96&q=75",
      "altText": "Clerk docs",
      "href": "https://clerk.com/docs"
    },
    "contact": {
      "email": "support@clerk.com",
      "name": "Clerk Platform Team",
      "url": "https://clerk.com/support"
    },
    "description": "The Clerk REST Platform API, for programmatic management of resources of a workspace.\n\n> [!note]\n> This API is currently in private beta. If you'd like access for your workspace, please fill out and submit our [request form](https://clerkdev.notion.site/2df2b9ab44fe803f9031e9f3185c5e19).",
    "version": "BETA",
    "termsOfService": "https://clerk.com/terms",
    "license": {
      "name": "MIT",
      "url": "https://github.com/clerk/openapi-specs/blob/main/LICENSE"
    }
  },
  "servers": [
    {
      "url": "https://api.clerk.com/v1"
    }
  ],
  "security": [
    {
      "platform_api_access_token": []
    }
  ],
...
  "paths": {
...
    "/platform/applications/{applicationID}/transfers": {
      "post": {
        "operationId": "PlatformCreateApplicationTransfer",
        "x-speakeasy-group": "platform",
        "x-speakeasy-name-override": "createApplicationTransfer",
        "tags": ["Application Transfers"],
        "summary": "Create an application transfer",
        "description": "Create a new transfer request for an application. This initiates the process of\ntransferring ownership of the application to another workspace.\n\nOnly one pending transfer can exist for an application at a time. The transfer\nwill expire after 24 hours if not completed.\n",
        "security": [
          {
            "platform_api_access_token": []
          }
        ],
        "parameters": [
          {
            "name": "applicationID",
            "in": "path",
            "description": "Application ID.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Application transfer created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformApplicationTransferResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/AuthenticationInvalid"
          },
          "403": {
            "$ref": "#/components/responses/AuthorizationInvalid"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "platform_api_access_token": {
        "type": "http",
        "scheme": "bearer",
        "description": "Platform API access token."
      }
    },
    "schemas": {
...
      "ClerkError": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "long_message": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "meta": {
            "type": "object"
          }
        },
        "required": ["message", "long_message", "code"]
      },
      "ClerkErrors": {
        "type": "object",
        "properties": {
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ClerkError"
            }
          },
          "meta": {
            "type": "object"
          },
          "clerk_trace_id": {
            "type": "string"
          }
        },
        "required": ["errors"]
      },
...
      "PlatformApplicationTransferResponse": {
        "title": "Application Transfer",
        "description": "Represents an application transfer request. Application transfers allow transferring ownership of an application from one organization to another.",
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "object": {
            "type": "string",
            "enum": ["application_transfer"],
            "description": "The object type."
          },
          "id": {
            "type": "string",
            "description": "The unique identifier for the application transfer.",
            "example": "appxfr_2abc123def456"
          },
          "code": {
            "type": "string",
            "description": "A unique code for the transfer that can be shared with the recipient to claim the application.",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "application_id": {
            "type": "string",
            "description": "The ID of the application being transferred.",
            "example": "app_2abc123def456"
          },
          "status": {
            "type": "string",
            "description": "The current status of the transfer.",
            "enum": ["pending", "completed", "canceled", "expired"],
            "example": "pending"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "description": "The timestamp when the transfer expires if not completed.",
            "example": "2024-01-16T12:00:00Z"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "The timestamp when the transfer was created.",
            "example": "2024-01-15T12:00:00Z"
          },
          "canceled_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "The timestamp when the transfer was canceled, or null if not canceled.",
            "example": null
          },
          "completed_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "The timestamp when the transfer was completed, or null if not completed.",
            "example": null
          }
        },
        "required": [
          "object",
          "id",
          "code",
          "application_id",
          "status",
          "expires_at",
          "created_at",
          "canceled_at",
          "completed_at"
        ]
      }
    },
    "responses": {
...
      "AuthenticationInvalid": {
        "description": "Authentication invalid",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ClerkErrors"
            }
          }
        }
      },
      "AuthorizationInvalid": {
        "description": "Authorization invalid",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ClerkErrors"
            }
          }
        }
      },
...
      "Conflict": {
        "description": "Conflict",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ClerkErrors"
            }
          }
        }
      }
    }
  },
...
}